+++ src/main/java/meal/planner/dataBase/itemList/ItemList.java
@@ -52,12 +52,12 @@ public class ItemList<T extends Item, S> {
 	}
 	
 	/** Returns the item with the given id */
-	public T findId(int id){
+	public T findId(long id) {
 		return findId(id, items);
 	}
 	
 	/** Returns the item with the given id */
-	private T findId(int id, ArrayList<T> items){
+	private T findId(long id, ArrayList<T> items) {
 		for(T t : items){
 			if(t.getId()==id){
 				return t;
